refactor:🌈 SceneDelegate에서 국가정보 setup 하는 로직 분리
- persistenceManager로 분리함으로써 SceneDelegate에서 Flagkit import할 필요 없어짐
- 또한 setupCountries, filterCountries 등 함수를 SceneDelegate에서 제거
- createCountriesWithAPIRequest 함수는 비동기처리하되 SceneDelegate에서 분리되었으므로
  main queue에 넣어줄 필요가 없어짐
